pyproject.toml
[project]
name = "homelab-sre-assistant"
version = "0.1.0"
description = "AI-powered SRE assistant for homelab infrastructure"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3",
    "langchain-chroma>=0.2",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "httpx>=0.28",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "langchain-openai>=1.1.9",
    "streamlit>=1.40",
    "prometheus-client>=0.21",
    "apscheduler>=3.10,<4",
]

[dependency-groups]
dev = [
    "mypy>=1.0",
    "ruff>=0.9",
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "respx>=0.22.0",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.mypy]
strict = true
python_version = "3.13"

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: tests that mock external HTTP APIs (run by default)",
    "e2e: tests that hit real services and cost money (skipped by default, run with --run-e2e)",
]
